create WidgetStatePigeon
